backend/app/services/google_calendar_service.py
from google.auth.transport.requests import Request
from google.oauth2.credentials import Credentials
from google_auth_oauthlib.flow import InstalledAppFlow
from googleapiclient.discovery import build
from googleapiclient.errors import HttpError
import os

class GoogleCalendarService:
    """Service for interacting with Google Calendar API."""
    
    # If modifying these scopes, delete the file token.json.
    SCOPES = ['https://www.googleapis.com/auth/calendar.readonly']
    
    def __init__(self):
        self.service = None
        self.credentials = None
    
    def authenticate(self, user_id: str):
        """Authenticate with Google Calendar API."""
        # TODO: Implement Google Calendar authentication
        # This will handle the OAuth flow and store credentials
        pass
    
    def get_events(self, user_id: str, start_time: str, end_time: str):
        """Fetch calendar events for a user within a time range."""
        # TODO: Implement calendar events retrieval
        # This will:
        # 1. Load user's stored credentials
        # 2. Build the Calendar API service
        # 3. Fetch events within the specified time range
        # 4. Return formatted event data
        
        # Mock data for now
        return [
            {
                "id": "event-1",
                "summary": "Team Standup",
                "start": {"dateTime": "2024-01-15T10:00:00-08:00"},
                "end": {"dateTime": "2024-01-15T10:30:00-08:00"},
                "location": "Conference Room A"
            },
            {
                "id": "event-2", 
                "summary": "Client Presentation",
                "start": {"dateTime": "2024-01-15T14:00:00-08:00"},
                "end": {"dateTime": "2024-01-15T15:00:00-08:00"},
                "location": "Zoom Meeting"
            }
        ]
    
    def revoke_access(self, user_id: str):
        """Revoke access to Google Calendar."""
        # TODO: Implement access revocation
        pass