backend/app/services/news_service.py
import requests
from bs4 import BeautifulSoup
from typing import List, Dict, Any
from app.core.config import settings

class NewsService:
    """Service for fetching news articles."""
    
    def __init__(self):
        self.api_key = settings.NEWS_API_KEY
        self.base_url = "https://newsapi.org/v2"
    
    def get_top_headlines(self, sources: List[str] = None, category: str = "general") -> List[Dict[str, Any]]:
        """Get top headlines from news sources."""
        # TODO: Implement actual news API call
        # This will make a request to News API
        
        # Mock data for now
        return [
            {
                "title": "Tech Stocks Surge on AI Breakthrough",
                "description": "Major technology stocks rose sharply following news of a significant artificial intelligence breakthrough.",
                "url": "https://example.com/tech-stocks",
                "published_at": "2024-01-15T08:00:00Z",
                "source": "Tech News Daily"
            },
            {
                "title": "Renewable Energy Research Shows Promise",
                "description": "Scientists announce a new solar panel technology that could revolutionize clean energy production.",
                "url": "https://example.com/renewable-energy",
                "published_at": "2024-01-15T07:30:00Z",
                "source": "Science Today"
            }
        ]
    
    def get_news_by_category(self, category: str, limit: int = 10) -> List[Dict[str, Any]]:
        """Get news articles by category."""
        # TODO: Implement category-based news fetching
        return self.get_top_headlines(category=category)
    
    def search_news(self, query: str, limit: int = 10) -> List[Dict[str, Any]]:
        """Search for news articles by query."""
        # TODO: Implement news search functionality
        return []