backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
langchain==0.0.350
langgraph==0.0.21
openai==1.3.7
beautifulsoup4==4.12.2
requests==2.31.0
google-api-python-client==2.110.0
google-auth-oauthlib==1.1.0
python-multipart==0.0.6